Rename module folder
